Move acrManager to container folder
